requirements.txt
tensorflow==2.13.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
augmentor==0.2.12
streamlit==1.28.1
pillow==10.0.0
